requirements.txt
alembic==1.10.2
aniso8601==9.0.1
apispec==5.2.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.11.2
blinker==1.4
Bootstrap-Flask==2.0.2
cachelib==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.0.1
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.0
Flask==2.1.1
Flask-AppBuilder==4.3.1
Flask-Assets==2.0
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Limiter==2.4.5.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Minify==0.39
flask-restx==1.1.0
Flask-SeaSurf==1.1.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.6.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==1.1.2
grpcio==1.51.3
grpcio-status==1.51.3
htmlmin==0.1.12
httplib2==0.21.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
is-safe-url==1.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsmin==3.0.1
jsonschema==4.17.3
lesscpy==0.15.0
limits==2.6.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
ordered-set==4.1.0
packaging==21.3
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
ply==3.11
prison==0.2.1
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.8
pyrsistent==0.19.3
pyScss==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.7.1
PyYAML==6.0
rcssmin==1.1.0
redis==4.5.1
requests==2.28.2
rich==12.3.0
rsa==4.9
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.40.0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.14
uWSGI==2.0.21
validate-email-address==1
webassets==2.0
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
xxhash==3.0.0
zipp==3.15.0
